requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-tinymce==3.7.1
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.13.3
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-tinymce==3.7.1
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.13.3
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1